Server/requirements.txt
aniso8601==8.0.0
beautifultable==1.0.1
bleach==5.0.0
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.12
click==7.1.2
commonmark==0.9.1
cryptography==38.0.3
docutils==0.18.1
Flask==1.1.2
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.3
idna==3.3
importlib-metadata==4.11.3
itsdangerous==1.1.0
Jinja2==2.11.2
keyring==23.5.0
MarkupSafe==1.1.1
pkginfo==1.8.2
pycparser==2.21
pycryptodome==3.15.0
Pygments==2.12.0
pytz==2020.1
pywin32-ctypes==0.2.0
readme-renderer==35.0
requests==2.27.1
requests-toolbelt==0.9.1
rfc3986==2.0.0
rich==12.3.0
six==1.15.0
SQLAlchemy==1.3.18
twine==4.0.0
urllib3==1.26.9
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
zipp==3.8.0